Mhd amen

+++ Modules/HousingMovements/Http/Controllers/CarTypeController.php
@@ -43,42 +43,61 @@ class CarTypeController extends Controller
                 ->editColumn('name_en', function ($row) {
                     return $row->name_en ?? '';
                 })
-                ->addColumn('action', function ($row) {
-                    $html = '';
-                    $html = '<div class="btn-group" role="group">
-                    <button id="btnGroupDrop1" type="button"
-                        style="background-color: transparent;
-                    font-size: x-large;
-                    padding: 0px 20px;"
-                        class="btn btn-secondary dropdown-toggle" data-toggle="dropdown"
-                        aria-haspopup="true" aria-expanded="false">
-                        <i class="fa fa-cog" aria-hidden="true"></i>
-                    </button>
-                    <div class="dropdown-menu">
-                        <a class="dropdown-item btn-modal" style="margin: 2px;"
-                            title="تعديل"
-                            href="' . route('cartype.edit', ['id' => $row->id]) . ' "
-                            data-href="' . route('cartype.edit', ['id' => $row->id]) . ' "
-                            data-container="#edit_car_type_model">
+                ->addColumn(
+                    'action',
+                    function ($row) {
 
-                            <i class="fas fa-edit cursor-pointer"
-                                style="padding: 2px;color:rgb(8, 158, 16);"></i>
-                            تعديل </a>
+                        $html = '';
 
-                        <a class="dropdown-item btn-modal" style="margin: 2px;" 
-                            href=" ' . route('cartype.delete', ['id' => $row->id]) . ' "
-                            data-href="' . route('cartype.delete', ['id' => $row->id]) . '"
-                            {{-- data-target="#active_auto_migration" data-toggle="modal" --}} {{-- id="delete_auto_migration" --}}>
+                        $html .= '
+                        <a href="' . route('cartype.edit', ['id' => $row->id])  . '"
+                        data-href="' . route('cartype.edit', ['id' => $row->id])  . ' "
+                         class="btn btn-xs btn-modal btn-info edit_carType_button"  data-container="#edit_car_type_model"><i class="fas fa-edit cursor-pointer"></i>' . __("messages.edit") . '</a>
+                    ';
+                        $html .= '
+                    <button data-href="' .  route('cartype.delete', ['id' => $row->id]) . '" class="btn btn-xs btn-danger delete_carType_button"><i class="glyphicon glyphicon-trash"></i>' . __("messages.delete") . '</button>
+                ';
 
-                            <i class="fa fa-trash cursor-pointer"
-                                style="padding: 2px;color:red;"></i>
-                            حذف
 
-                        </a>
-                    </div>
-                </div>';
-                    return $html;
-                })
+                        return $html;
+                    }
+                )
+                // ->addColumn('action', function ($row) {
+                //     $html = '';
+                //     $html = '<div class="btn-group" role="group">
+                //     <button id="btnGroupDrop1" type="button"
+                //         style="background-color: transparent;
+                //     font-size: x-large;
+                //     padding: 0px 20px;"
+                //         class="btn btn-secondary dropdown-toggle" data-toggle="dropdown"
+                //         aria-haspopup="true" aria-expanded="false">
+                //         <i class="fa fa-cog" aria-hidden="true"></i>
+                //     </button>
+                //     <div class="dropdown-menu">
+                //         <a class="dropdown-item btn-modal" style="margin: 2px;"
+                //             title="تعديل"
+                //             href="' . route('cartype.edit', ['id' => $row->id]) . ' "
+                //             data-href="' . route('cartype.edit', ['id' => $row->id]) . ' "
+                //             data-container="#edit_car_type_model">
+
+                //             <i class="fas fa-edit cursor-pointer"
+                //                 style="padding: 2px;color:rgb(8, 158, 16);"></i>
+                //             تعديل </a>
+
+                //         <a class="dropdown-item btn-modal" style="margin: 2px;" 
+                //             href=" ' . route('cartype.delete', ['id' => $row->id]) . ' "
+                //             data-href="' . route('cartype.delete', ['id' => $row->id]) . '"
+                //             {{-- data-target="#active_auto_migration" data-toggle="modal" --}} {{-- id="delete_auto_migration" --}}>
+
+                //             <i class="fa fa-trash cursor-pointer"
+                //                 style="padding: 2px;color:red;"></i>
+                //             حذف
+
+                //         </a>
+                //     </div>
+                // </div>';
+                //     return $html;
+                // })
                 ->filter(function ($query) use ($request) {
 
                     // if (!empty($request->input('full_name'))) {
@@ -185,11 +204,18 @@ class CarTypeController extends Controller
      */
     public function destroy($id)
     {
-        try {
-            CarType::find($id)->delete();
-            return redirect()->back()->with(__('deleted_success'));
-        } catch (Exception $e) {
-            return redirect()->back();
+       
+        if (request()->ajax()) {
+            try {
+                CarType::find($id)->delete();
+                $output = [
+                    'success' => true,
+                    'msg' => 'تم حذف نوع السيارة بنجاح',
+                ];
+            } catch (Exception $e) {
+                return redirect()->back();
+            }
+            return $output;
         }
     }
